Added Neural Network for chi squared selected features
I added a model with a Neural Network for the chi squared selected features. In the first section (3a1) we do some hyperparameter tuning to find out the best values for the number of neurons in each layer, the maximum number of iterations and the learning rate (this hyperparameter tuning can take around 3 minutes if you run it). In the second section (3a2) we use a model with the hyperparameters we just computed, and we fit the model. Finally in the third section we perform a k-fold cross validation.